Modified the check_expired_files to save the new chat history into the disk while also retunring it.
When the 'check_expired_files' would get the chat_history, it initially didn't save back the new chat_history in the disk, but instead forwarded it towards the model for response. 

This was fine up until some error would occur during conversation. If there are files that needed to be removed that are mentioned in the times_files.json. It would empty the json file and return the chat history. But now as this particular object is not stored in the disk, the old object remains in the disk. 
And now if an error occurs with the new chat history, then it won't be saved into the disk and the old chat history with the files will remain even though the json file would be empty.